+++ src/sf/ui/window.cpp
@@ -34,8 +34,8 @@ void Window::render(ConsoleKey key) {
   }
   if (this->window_title.length() > 0) {
     size_t start_x =
-        round(this->width / 2.0f - this->window_title.length() / 2.0f);
-    mvwprintw(this->win, 0, start_x, this->window_title.c_str());
+        (size_t)round(this->width / 2.0f - this->window_title.length() / 2.0f);
+    mvwprintw(this->win, 0, (int)start_x, this->window_title.c_str());
   }
 
   wrefresh(this->win);
@@ -51,8 +51,8 @@ void Window::calcWindowSize(int &x, int &y, int &width, int &height) {
   x = 0;
   y = 0;
 
-  width = round(COLS * this->size_x);
-  height = round(LINES * this->size_y);
+  width = (int)round(COLS * this->size_x);
+  height = (int)round(LINES * this->size_y);
 
   if (this->alignment_x == WindowAlignment::RIGHT) {
     x = COLS - width;
